+++ src/viewers/dir_viewers.py
@@ -1,9 +1,9 @@
 from PySide6.QtWidgets import QWidget,QFileSystemModel,QVBoxLayout,QTableView,QListView,QSizePolicy,QPushButton
 from PySide6.QtCore import QSize
-from src.functions import GUIFunctions
 class DirViewers(QWidget):
     def __init__(self,parent = None):
         super().__init__()
+        
         self.file_system_model = QFileSystemModel()
         self.file_system_model.setRootPath("")
         self.list_view = QListView(self)
@@ -14,7 +14,6 @@ class DirViewers(QWidget):
         self.list_view.setGridSize(QSize(100, 100))
         self.list_view.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
         self.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-        self.gui = GUIFunctions(self)
         self.uper_path = ""
 
         layout = QVBoxLayout(self)
@@ -25,10 +24,18 @@ class DirViewers(QWidget):
         self.file_system_model.setRootPath(dir_path)
         self.list_view.setRootIndex(self.file_system_model.index(dir_path))
 
-    def itme_list_clicked(self,index):
+    def itme_list_clicked(self,index,parent):
+        """Obsługuje kliknięcie elementu – jeśli katalog, przechodzi do niego."""
         file_path = self.file_system_model.filePath(index)
-        print(f"Clicked item path: {file_path}")
-        self.gui.display_file_content(file_path)
+        
+        if self.file_system_model.isDir(index):  
+            print(f"Przechodzę do katalogu: {file_path}")
+            self.set_directory(file_path)
+        else:
+            import src.gui_function as g
+            g.display_file_content(parent,file_path)
+            print(f"Kliknięto plik: {file_path}")
+        
 def display_dir_content(context,dir_path):
     dir_viewers = DirViewers()
     dir_viewers.file_system_model.setRootPath(dir_path)
@@ -37,7 +44,7 @@ def display_dir_content(context,dir_path):
     prev_btn = QPushButton("<-")
     next_btn = QPushButton("->")
 
-    dir_viewers.list_view.clicked.connect(dir_viewers.itme_list_clicked)
+    dir_viewers.list_view.clicked.connect(lambda index: dir_viewers.itme_list_clicked(index, context))
     
     layout = context.ui.reportsPage.layout()
     if layout is None:
